ensure that values are logged as unsigned numbers
git-svn-id: http://arthurdejong.org/svn/nss-pam-ldapd/nss-pam-ldapd@1838 ef36b2f9-881f-0410-afb5-c4e39611909c

+++ nslcd/group.c
@@ -335,7 +335,7 @@ NSLCD_HANDLE(
   gid_t gid;
   char filter[4096];
   READ_TYPE(fp,gid,gid_t);
-  log_setrequest("group=%d",(int)gid);,
+  log_setrequest("group=%lu",(unsigned long int)gid);,
   NSLCD_ACTION_GROUP_BYGID,
   mkfilter_group_bygid(gid,filter,sizeof(filter)),
   write_group(fp,entry,NULL,&gid,1,session)
